:truck: Rename(mainPage): 컴포넌트 위치 수정
Related to: #56
