+++ src/app/(main)/_components/activity-nav-bar.tsx
